+++ site-new/src/page-components/community/index.tsx
@@ -1,7 +1,7 @@
 import Hero from "@site/src/components/Hero";
 import Background from "@site/src/components/Background";
 import { TextIconFeature } from "@site/src/components/TextIconFeature";
-import TextIconCard from "@site/src/components/TextIconCard";
+import Card from "@site/src/components/Card";
 import { MarqueeTextSkewed } from "@site/src/components/Marquee";
 
 const features = [
@@ -97,7 +97,7 @@ function Community() {
       <div className="container mx-auto px-4 py-12">
         <div className="grid grid-cols-1 gap-8 md:grid-cols-2 lg:grid-cols-3">
           {explore.map((feature, index) => (
-            <TextIconCard
+            <Card
               key={index}
               title={feature.title}
               text={feature.text}
