--- site-new/src/pages/component-guide/image-card.mdx
@@ -1,76 +1,0 @@
-import React from "react";
-import ImageCard from "../../components/ImageCard";
-
-## ImageCard Component
-
-<div className="component-guide text-white">
-<p>The ImageCard component is used to display an image alongside some text and a button, with configurable orientation and size. It is useful for displaying promotional content or feature highlights with a visually appealing layout.</p>
-
-## Properties
-
-| Property    | Type   | Required | Default    | Description                                                                                     |
-| ----------- | ------ | -------- | ---------- | ----------------------------------------------------------------------------------------------- |
-| title       | string | Yes      | ""         | The title text displayed in the card.                                                           |
-| image       | string | No       | NoImageUrl | The URL of the image displayed in the card. If not provided, a default image is shown.          |
-| text        | string | Yes      | ""         | The main body text displayed alongside the image.                                               |
-| buttonText  | string | Yes      | ""         | The text displayed on the button.                                                               |
-| url         | string | Yes      | ""         | The URL that the button will link to.                                                           |
-| orientation | string | No       | "vertical" | The layout orientation of the card, either "horizontal" or "vertical".                          |
-| size        | string | No       | "large"    | The size of the card, can be "large", "medium", or "small". Determines the image and text size. |
-
-## Usage
-
-```jsx
-<ImageCard
-  title="Sample Title"
-  image="https://example.com/image.png"
-  text="This is the body text for the card."
-  buttonText="Click Me"
-  url="/sample-page"
-  orientation="horizontal"
-  size="medium"
-/>
-```
-
-## Examples
-
-<p>Horizontal ImageCard</p>
-<ImageCard
-  buttonText="Button Text"
-  text="This is a text body"
-  title="This is the Title"
-  orientation="horizontal"
-  url="/"
-/>
-
-<p>Vertical ImageCard - Large</p>
-<ImageCard
-  buttonText="Button Text"
-  text="This is a text body"
-  title="This is the Title"
-  orientation="vertical"
-  url="/"
-  size="large"
-/>
-
-<p>Vertical ImageCard - Medium</p>
-<ImageCard
-  buttonText="Button Text"
-  text="This is a text body"
-  title="This is the Title"
-  orientation="vertical"
-  url="/"
-  size="medium"
-/>
-
-<p>Vertical ImageCard - Small</p>
-<ImageCard
-  buttonText="Button Text"
-  text="This is a text body"
-  title="This is the Title"
-  orientation="vertical"
-  url="/"
-  size="small"
-/>
-
-</div>
