+++ site-new/src/pages/index.tsx
@@ -5,7 +5,7 @@ import Background from "../components/Background";
 import Rocket from "@site/static/img/Rocket";
 import Underline from "@site/static/img/Underline";
 import Hero from "../components/HeroCard";
-import TextIconCard from "../components/TextIconCard";
+import Card from "../components/Card";
 import tbdRex from "@site/static/img/tbd-rex";
 import { PixelBorderWrapper } from "../components/PixelBorder";
 import { useEffect, useRef } from "react";
@@ -137,25 +137,25 @@ export default function Home(): JSX.Element {
         <Background bgColor="black" className="py-0 lg:py-24">
           <div className="container mx-auto px-4">
             <div className="grid grid-cols-1 gap-6 lg:grid-cols-2">
-              <TextIconCard
+              <Card
                 title="Open Protocols"
                 text="We’re building open protocols, standards, and libraries. Our projects are open source because we want everyone in the economy to benefit: individuals, businesses, institutions, and government."
                 theme="yellow"
                 icon={tbdRex}
               />
-              <TextIconCard
+              <Card
                 title="Open Libraries"
                 text="We’re building open protocols, standards, and libraries. Our projects are open source because we want everyone in the economy to benefit: individuals, businesses, institutions, and government."
                 theme="teal"
                 icon={tbdRex}
               />
-              <TextIconCard
+              <Card
                 title="Open Standards"
                 text="We’re building open protocols, standards, and libraries. Our projects are open source because we want everyone in the economy to benefit: individuals, businesses, institutions, and government."
                 theme="purple"
                 icon={tbdRex}
               />
-              <TextIconCard
+              <Card
                 title="Open Source"
                 text="We’re building open protocols, standards, and libraries. Our projects are open source because we want everyone in the economy to benefit: individuals, businesses, institutions, and government."
                 theme="yellow"
